updated layeredTexture, with testcase
layeredTexture has 4 arrays that are not supported by mtoer translator
at the momment.  the default arrays have 2 layers, a transparent green
layer over a red one . so it comes out a orange render result.  It
should be a green checker over a gray cloth(refer to the image done by
mentalray) after the mtoer translator works with the arrays.